pyproject.toml
[tool.poetry]
name = "nonebot-plugin-manga-translator"
version = "0.1.0"
description = "基于nonebot2的适配多种api的图片/漫画翻译插件"
authors = ["xenophon <674550338@qq.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "nonebot_plugin_manga_translator"}]

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.0.0rc1"
Pillow = "^9.1.1"
nonebot-adapter-onebot = "^2.1.0"
httpx = "^0.23.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
